Fixed interleave parameter in ReSID::executeCycles

+++ Emulator/SID/SIDBridge.cpp
@@ -473,7 +473,7 @@ SIDBridge::_dump(int nr)
     
     msg("ReSID:\n");
     msg("------\n");
-    msg("    Chip model: %d (%s)\n", (int)residRev, sidRevisionName(residRev));
+    msg("    Chip model: %ld (%s)\n", residRev, sidRevisionName(residRev));
     msg(" Sampling rate: %f\n", resid[nr].getSampleRate());
     msg(" CPU frequency: %d\n", resid[nr].getClockFrequency());
     msg("Emulate filter: %s\n", resid[nr].getAudioFilter() ? "yes" : "no");
